TTM-41 realisation 6

+++ app/src/main/java/com/example/tabletopapplication/presentationlayer/models/Material/MaterialDao.kt
@@ -20,5 +20,5 @@ interface MaterialDao {
     fun getAllMaterials(): LiveData<List<Material>>
 
     @Query("Select * from materialsTable Where id=:id")
-    fun getOneMaterial(id:Int): Material
+    fun getOneMaterial(id:Long): Material
 }
