+++ app/src/main/java/com/example/tabletopapplication/presentationlayer/models/Material/MaterialRepository.kt
@@ -7,7 +7,7 @@ class MaterialRepository(private val materialDao: MaterialDao) {
 
     val allMaterials: LiveData<List<Material>> = materialDao.getAllMaterials()
 
-    fun getOneMaterial(id: Int): Material {
+    fun getOneMaterial(id: Long): Material {
         return materialDao.getOneMaterial(id)
     }
